These branches split at 91f08ae2 in December 2021 and ran in parallel for
four and a half years in separate clones, roughly 400 commits on Legacy and
240 on dev, with neither clone holding the other's objects. This merge joins
them so no work is stranded on a branch nobody builds.

Authority was applied per area rather than per file. dev wins where it is
ahead: the annotation protos, the gRPC annotation service, the
transport-agnostic WebAnnotationModel and its WCF and gRPC implementations,
the AnnotationDatabase SSDT project, and the whole Identity Server stack.
Legacy wins everywhere else, which is most of the client: SDK-style projects,
MonoGame instead of XNA, the WPF login and preferences UI, Velopack, the
dependency-injection module loader, the texture pipeline, the Bajaj mesh
work, and the SAM2 segmentation gRPC stack that dev never had.

Of 433 conflicts, 306 resolved to Legacy and 99 to dev; the rest were
add/delete pairs where one side had intentionally removed a file. Deletions
Legacy made deliberately are preserved: the WinForms logon superseded by the
WPF window, the mesh generators superseded by Bajaj, and the packages.config
files dropped during the SDK migration.

The submodule sets were disjoint and are now unioned, so both the
segmentation protos and the NetTopologySuite sources are declared.

Clients/AnnotationModel is dropped as dead code, along with a stale duplicate
of the ConnectomeDataModelCore test project and a stray server log.

Everything.sln is carried over from Legacy unchanged here and rebuilt in the
following commit, because a hand-merge of two solution files this size is far
more error-prone than regenerating the project list.

+++ Servers/AnnotationService/AnnotationService/Service/Types/StructureTypeExtensions.cs
@@ -47,7 +47,7 @@ namespace AnnotationService.Types
             st.StructureTags = type.StructureTags is null ? [] : type.StructureTags.Split(';');
 
             st.Abstract = type.Abstract;
-            st.Color = type.Color;
+            st.Color = (uint)type.Color;
             st.Code = type.Code;
             st.HotKey = type.HotKey.Length > 0 ? type.HotKey[0] : '\0';
             st.PermittedLinks = PopulateLinks(type);
@@ -79,7 +79,7 @@ namespace AnnotationService.Types
 
             type.StructureTags = structuretags;
             type.Abstract = st.Abstract;
-            type.Color = st.Color;
+            type.Color = (int)st.Color;
             type.Code = st.Code;
             type.HotKey = st.HotKey.ToString();
             type.Username = Annotation.ServiceModelUtil.GetUserForCall();
